+++ source/collie/codec/lengthbaseframe.d
@@ -145,7 +145,7 @@ protected:
 		return data;
 	}
 	
-	void callBack(ubyte[] data,uint size)
+	void callBack(ubyte[] data,size_t size)
 	{}
 
 private:
@@ -198,7 +198,7 @@ unittest
 		{
 		}
 		
-		override void fireWrite(ubyte[] msg, void delegate(ubyte[], uint) cback = null)
+		override void fireWrite(ubyte[] msg, void delegate(ubyte[], size_t) cback = null)
 		{
 			gloaData ~= msg;
 			writeln("length is : ", msg[0 .. 4], " \n the data is : ", cast(string)(msg[4 .. $]));
